Former-commit-id: 631344e1f2517a511f0b3b877acc4e80a68c3ce5 [formerly 515910f81b988b3ff9df081d639fffd8df04bb7d] [formerly 6781a775127039b7e3fcd0f44d5fbde308dd17c4 [formerly 4690ef3819e0268f4aa8abf68e9b87fa3de4b15b]]
Former-commit-id: 9a89f053586566ee22cedfcea9c8a187256f4fc1 [formerly 67ec1747da393f855ba509fc21636bb3d91cb599]
Former-commit-id: 41f11195fa21427aa5e834f945b4d47509089c30

+++ src/Nazara/Utility/Win32/WindowImpl.cpp
@@ -394,7 +394,7 @@ namespace Nz
 	void WindowImpl::SetMaximumSize(int width, int height)
 	{
 		RECT rect = {0, 0, width, height};
-		AdjustWindowRect(&rect, GetWindowLongPtr(m_handle, GWL_STYLE), false);
+		AdjustWindowRect(&rect, static_cast<DWORD>(GetWindowLongPtr(m_handle, GWL_STYLE)), false);
 
 		if (width != -1)
 			m_maxSize.x = rect.right-rect.left;
@@ -410,7 +410,7 @@ namespace Nz
 	void WindowImpl::SetMinimumSize(int width, int height)
 	{
 		RECT rect = {0, 0, width, height};
-		AdjustWindowRect(&rect, GetWindowLongPtr(m_handle, GWL_STYLE), false);
+		AdjustWindowRect(&rect, static_cast<DWORD>(GetWindowLongPtr(m_handle, GWL_STYLE)), false);
 
 		if (width != -1)
 			m_minSize.x = rect.right-rect.left;
@@ -432,7 +432,7 @@ namespace Nz
 	{
 		// SetWindowPos demande la taille totale de la fenêtre
 		RECT rect = {0, 0, static_cast<LONG>(width), static_cast<LONG>(height)};
-		AdjustWindowRect(&rect, GetWindowLongPtr(m_handle, GWL_STYLE), false);
+		AdjustWindowRect(&rect, static_cast<DWORD>(GetWindowLongPtr(m_handle, GWL_STYLE)), false);
 
 		SetWindowPos(m_handle, nullptr, 0, 0, rect.right - rect.left, rect.bottom - rect.top, SWP_NOMOVE | SWP_NOZORDER);
 	}
